# apps/web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchLoadInput": {
        "additionalProperties": false,
        "properties": {
          "assessment_input_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Assessment Input Hash",
            "type": "string"
          }
        },
        "required": [
          "assessment_input_hash"
        ],
        "title": "ResearchLoadInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Reviso research API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/theses/{thesis_id}/research-context": {
      "post": {
        "operationId": "load_research_theses__thesis_id__research_context_post",
        "parameters": [
          {
            "in": "path",
            "name": "thesis_id",
            "required": true,
            "schema": {
              "title": "Thesis Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchLoadInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Load Research Theses  Thesis Id  Research Context Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Research"
      }
    },
...
  }
}